requirements.txt
python-telegram-bot==20.5
pandas==2.1.0
matplotlib==3.7.3
python-binance==1.0.19
python-dotenv==1.0.0
requests==2.31.0
alpha_vantage==2.3.1